requirements.txt
accelerate==1.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
diffusers==0.32.1
einops==0.8.1
fairscale==0.4.13
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.4
loralib==0.1.2
lpips==0.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
munch==4.0.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
peft==0.15.2
pillow==11.1.0
protobuf==6.30.0
psutil==7.0.0
PyYAML==6.0.2
qwen-vl-utils[decord]==0.0.8
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.49.0
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.3.0
zipp==3.21.0
